requirements.txt
streamlit
pandas
numpy
joblib
scikit-learn
fastapi
uvicorn
requests
psycopg2-binary
tensorflow
nltk
pydantic
SQLAlchemy>=1.4,<2.0
passlib[bcrypt]
python-jose
passlib
langchain_groq
langchain
transformers
beautifulsoup4
torch
googlesearch-python
streamlit-cookies-manager